+++ noodles-sam/src/header/parser/record/value/comment.rs
@@ -2,8 +2,6 @@ use std::{error, fmt};
 
 use bstr::BString;
 
-use crate::header::parser::record::split_off_first;
-
 /// An error returned when a SAM header record comment value fails to parse.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ParseError {
@@ -31,7 +29,7 @@ pub(super) fn parse_comment(src: &mut &[u8]) -> Result<BString, ParseError> {
 fn consume_delimiter(src: &mut &[u8]) -> Result<(), ParseError> {
     const PREFIX: u8 = b'\t';
 
-    if let Some(&PREFIX) = split_off_first(src) {
+    if let Some(&PREFIX) = src.split_off_first() {
         Ok(())
     } else {
         Err(ParseError::InvalidDelimiter)