+++ noodles-vcf/src/io/reader/record_buf/info/field/value.rs
@@ -112,10 +112,10 @@ fn parse_flag(s: &str) -> Result<Value, ParseError> {
 fn parse_raw_char(s: &str) -> Result<char, ParseError> {
     let mut chars = s.chars();
 
-    if let Some(c) = chars.next() {
-        if chars.next().is_none() {
-            return Ok(c);
-        }
+    if let Some(c) = chars.next()
+        && chars.next().is_none()
+    {
+        return Ok(c);
     }
 
     Err(ParseError::InvalidCharacter)