+++ noodles-bcf/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-  * bcf: Raise minimum supported Rust version (MSRV) to 1.87.0.
+  * bcf: Raise minimum supported Rust version (MSRV) to 1.88.0.
 
 ## 0.81.0 - 2025-12-18